Last Successful Version sent by Marijana
Marijana: Code is prepared for the 3 sensors, and is currently only
printing the result on the Netbeans terminal. But in future this(output
from sensors) will be direclty send to the Context server (pearsons ip).
The idea is to have Pearsons computer as second server, which will
present the Visual user interface (after getting sensor info from
Interpreter server).